option to restore

+++ js/apps/mail/cmps/send-mail.cmp.js
@@ -3,7 +3,7 @@ import { mailService } from "../services/mail-service.js";
 
 export default {
     name: 'send-mail',
-    props: [],
+    props: ['draft'],
     template: `
         <section class="send-mail">
         <form @submit.prevent="send" >
@@ -13,6 +13,7 @@ export default {
             </div>
            
             <div class="body">
+                {{draft}}
                     <input v-model=message.to type="email"  placeholder="To" required>
                     <input v-model=message.subject type="text" placeholder="Subject">
                     <textarea v-model=message.body name="body" id="body" cols="30" rows="14" placeholder="Your message">
@@ -20,9 +21,9 @@ export default {
                 <div class="send-btn-container">
                     <button class="send">send</button>
                     <button class="trash" @click="removeMail" title="Trash" ><i class="fa-solid fa-trash-can"></i></button>
-                </form>
                 </div>
             </div>
+        </form>
         </section>
     `,
     components: {
@@ -33,30 +34,25 @@ export default {
                 to: '',
                 subject: '',
                 body: '',
-
-                // id: '',
-                // subject,
-                // body,
-                // sentAt,
-                // to,
-                // isRead,
-                // isInbox,
-                // isStarred: false,
-                // isTrashed: false,
-                // isDraft,
             },
             interval: null,
 
         }
     },
     created() {
-        mailService.createDraft(this.message)
-            .then(mail => {
-                mail.isTrashed = true
-                this.message = mail
-                console.log(this.message);
-            })
-
+        if((Object.keys(this.draft).length === 0)){
+            console.log( this.draft);
+            mailService.createDraft(this.message)
+                .then(mail => {
+                    mail.isTrashed = true
+                    this.message = mail
+                    console.log(this.message);
+                })
+        }else{
+            this.message=this.draft
+            console.log( this.message);
+        }
+        
         this.interval = setInterval(() => {
             mailService.save(this.message)
             console.log('save?');
@@ -67,11 +63,10 @@ export default {
         String.prototype.isEmpty = function () {
             return (this.length === 0 || !this.trim());
         };
-        console.log(' '.isEmpty());
-        if (this.message.to.isEmpty() || this.message.subject.isEmpty() || this.message.body.isEmpty) {
+        
+        if (this.message.to.isEmpty() && this.message.subject.isEmpty() && this.message.body.isEmpty()) {
             mailService.remove(this.message.id)
         }
-        // console.log(' '.isEmpty());
         clearInterval(this.interval)
     },
     methods: {
@@ -87,7 +82,6 @@ export default {
             console.log(this.message.id);
             mailService.remove(this.message.id)
                 .then(() => {
-                    // this.$emit('opened')
                     this.close()
                 })
         },
